requirements.txt
bcrypt==3.2.2
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
protobuf==3.20.1
pycparser==2.21
PyMySQL==1.0.2
SQLAlchemy==1.4.40
Werkzeug==2.2.2
WTForms==3.0.1
